Organized and cleaned up Code
